+++ widgets/ForwardWidgets-huangxd/widgets/live.js
@@ -201,7 +201,7 @@ WidgetMetadata = {
             ],
         },
     ],
-    version: "1.0.8",
+    version: "1.0.9",
     requiredVersion: "0.0.1",
     description: "解析直播订阅链接【五折码：CHEAP.5;七折码：CHEAP】",
     author: "huangxd",
@@ -399,6 +399,7 @@ function parseM3UContent(content, iconList, bgColor, direction) {
                 backdropPath: backdropIcon || currentItem.cover || "https://i.miji.bid/2025/05/17/c4a0703b68a4d2313a27937d82b72b6a.png",
                 previewUrl: "", // 直播通常没有预览URL
                 link: url,
+                playerType: "system",
                 // 额外的元数据
                 metadata: {
                     group: currentItem.group,
@@ -448,6 +449,7 @@ async function loadDetail(link) {
                 posterPath: "https://i.miji.bid/2025/05/17/561121fb0ba6071d4070627d187b668b.png",
                 backdropPath: "https://i.miji.bid/2025/05/17/561121fb0ba6071d4070627d187b668b.png",
                 link: videoUrl,
+                playerType: "system",
             };
             childItems = [hint_item]
         }
@@ -461,6 +463,7 @@ async function loadDetail(link) {
             "Referer": link,
             "User-Agent": "AptvPlayer/1.4.6",
         },
+        playerType: "system",
         childItems: childItems,
     };