+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-15 03:14:22
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 383b2edd7b74aa4a93dccd7fe4a71fe6b13ef228
+
+## [ForwardWidgets-2kuai] - 2025-08-15 03:14:22
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 8b2af3d539b25a395a55cf9a91aa19072695acfd
+
+## [ForwardWidgets-pack1r] - 2025-08-15 03:14:22
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: e53c879bbeeb935307b15eda45d3b25c207af078
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-14 03:13:19
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main